+++ docxbox/docx/component/numbering.cc
@@ -3,7 +3,6 @@
 
 #include <docxbox/docx/component/numbering.h>
 
-// TODO(kay): transform into docx_renderer_numbering
 numbering::numbering(std::string path_extract) {
   path_extract_ = std::move(path_extract);
 }
@@ -60,7 +59,7 @@ std::string numbering::RenderXml(bool unordered) {
 
   return
       RenderXmlNamespaceTag() +
-      //RenderXmlAbstractNum0() +
+//      RenderXmlAbstractNum0() +
       "<w:abstractNum w:abstractNumId=\"2\">"
       "<w:multiLevelType w:val=\"hybridMultilevel\"/>"
       "<w:name w:val=\"Numbered list 1\"/>"
@@ -173,7 +172,7 @@ std::string numbering::RenderXml(bool unordered) {
       "</w:pPr>"
       "</w:lvl>"
       "</w:abstractNum>"
-      
+
       "<w:num w:numId=\"3\">"
         "<w:abstractNumId w:val=\"1\"/>"
       "</w:num>"