+++ CHANGELOG.md
@@ -3,7 +3,12 @@ docxBox Changelog
 
 V0.0.5 - Not finished yet
 -------------------------
+* Add batch sequence processing
 * Colorize side-by-side DOCX comparison (ls) to indicate additions/removals
+* Add replace text by run with text
+* Add replace text by paragraph containing run with text
+* Extend rendering of generic content types: render optional prefix and postfix 
+* Add insert paragraph and prefix/postfix
 
 V0.0.4 - 2020-05-26
 -------------------